Refactor AlterShardingTableRuleStatementUpdaterTest
